refactor: mailer job organization
